dcache-namespace: add delete to the namespace request
Added delete to the list of requests that can be performed
with the dcache-namespace element. With this delete method,
an AJAX request will be send to the specified url which must
correspond to a resource that can be deleted.

For dcache, this resources is a file, below is a simple usage
of how to use dcache-namespace element to delete a file:

```javascript
	var ns = document.createElement('dcache-namespace');
	var url = <url path to the file>;
	ns.auth = <supply credential>;
	ns.delete({
            url: url
        });
```