catch error before request is send
Modification:

0. update the precondition.js to version 0.0.2 in bower.json.
1. change the link tag to script tag for precondition.js
2. make all ajax request except `ls` to require both url and path
3. create a function (called `_appendFullPathToUrl`) to concatenate
the url and path. This function also ensure that both the url and
path are not null nor empty by using the precondition `checkArgument`
method.
4. use `_appendFullPathToUrl` function in all (except `ls`) request
and check any other argument necessary.

Result:

Proper handling of request and necessary error message will be display
even before the request is send.